+++ pit_components/lib/components/adv_list_view.dart
@@ -82,7 +82,7 @@ class AdvListView extends StatelessWidget {
       }
     }
 
-    if (this.onlyInner) newChildren.removeAt(newChildren.length - 1);
+    if (this.onlyInner && newChildren.length > 0) newChildren.removeAt(newChildren.length - 1);
 
     return newChildren;
   }