+++ pit_components/lib/components/adv_row.dart
@@ -72,7 +72,7 @@ class AdvRow extends StatelessWidget {
       }
     }
 
-    if (onlyInner) newChildren.removeAt(newChildren.length - 1);
+    if (onlyInner && newChildren.length > 0) newChildren.removeAt(newChildren.length - 1);
 
     return newChildren;
   }