+++ pit_components/lib/components/adv_state.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 abstract class AdvState<T extends StatefulWidget> extends State<T> {
   bool _firstRun = true;
+  bool _processing = false;
 
   @override
   Widget build(BuildContext context) {
@@ -16,4 +17,16 @@ abstract class AdvState<T extends StatefulWidget> extends State<T> {
   void initStateWithContext(BuildContext context) {}
 
   Widget advBuild(BuildContext context);
+
+  bool isProcessing() => _processing;
+
+  void process(Function f) async {
+    setState(() {
+      _processing = true;
+    });
+    await f();
+    setState(() {
+      _processing = false;
+    });
+  }
 }