+++ pit_components/lib/components/adv_text_field_plain.dart
@@ -141,8 +141,8 @@ class _AdvTextFieldPlainState extends State<AdvTextFieldPlain>
 
   _updateEffectiveSelection() {
     if (_textEditingCtrl.selection != _effectiveController.selection &&
-        _textEditingCtrl.selection.start <= _effectiveController.text.length &&
-        _textEditingCtrl.selection.end <= _effectiveController.text.length) {
+        _textEditingCtrl.selection.start <= (_effectiveController.text?.length ?? 0) &&
+        _textEditingCtrl.selection.end <= (_effectiveController.text?.length ?? 0)) {
       _effectiveController.removeListener(_update);
       _effectiveController.selection = _textEditingCtrl.selection;
       _effectiveController.addListener(_update);